Fix exclamation mark issue for dual data
As per the default design, the exclamation mark will be shown on both
SUBs when only DDS mobile data is disabled or DDS mobile network isn't
validated. To satisfy with dual data requirement, one change is
introduced to separate exclamation mark display for both SUBs, but
as long with baseline upgrade, the original change was broken, so
submitting this change is to fix exclamation issue based on dual data
requirement that needs to display exclamation marks separately.

Change-Id: I6b2dd363c4ef7e0bfaf8851ead789a34b3607cfb
CRs-Fixed: 3687337