change the initialization order for adding overlays
ScreenDecorations adds "overlay" views to the window hierarchy, which
house the appropriate decorations for various pieces of SystemUI. This
change simply swaps the order in which we initalize those overlays such
that the HWC overlay (special hardware-accelerated layers), if
available, are always added last and are thus at the highest z-index.

The reasoning for this is that some display pipelines are optimized to
only process a HWC layer if it is on top.

Note that this change means that any UI hosted in a software layer can
potentially be covered by the HWC layer which is now on-top. This
shouldn't be an issue in practice.

Test: on device with HWC support, launch camera app and verify from
SurfaceFlinger dumpsys that the HWC layers have the highest z-index
Bug: 313038036
Flag: NONE

Change-Id: I2e88774db0a40918c7f0be580ff66fe7a2093798
CRs-Fixed: 3657031
(cherry picked from commit 9ec91d5529f1523834b0a4142a988b5cdeb70340)