base: Fix stream value for Multi DP connections
Multi DPs work on the basis of steram id to distinguish b/w different connections. stream=0 or stream=1 should be assigned based on the connection availability. Current logic sets stream 1 to uevent info, if there is 1 connection existing. However, on disconnection stream values are not reset and the stream state remains 1. This causes stream=1 connection for single connection, which contradicts at
audio drivers.

Modify the logic to properly check available state. Keep a hashmap to use the first available state and remove the states on disconnection.

CRs-Fixed: 3804120

Change-Id: I6029cfc6c82c73c6eb3f3d2343dbe7e08941c81b